cleaned up the code to reduce bundle size, tried switching to bun or esbuild but ran into renaming issues.
