+++ src/core/rapier.js
@@ -1,0 +1,6 @@
+export let RAPIER = null
+
+import('@dimforge/rapier3d').then(rap => {
+    RAPIER = rap
+    document.dispatchEvent(new CustomEvent(`engine-started`, {bubbles: true}))
+})
