--- src/entities/developer/tools/DimensionTool.js
@@ -1,8 +1,0 @@
-import { Tool } from "./Tool.js";
-
-export class DimensionTool extends Tool {
-    constructor(direction){
-        super()
-    }
-
-}
