+++ src/interaction/hand.js
@@ -1,7 +1,8 @@
 import * as THREE from 'three'
+import { RAPIER } from '../core/rapier'
 import { XRControllerModelFactory } from 'three/addons/webxr/XRControllerModelFactory.js'
 import { XRHandModelFactory } from 'three/addons/webxr/XRHandModelFactory.js'
-import { RAPIER, INPUT_COLLIDER_HANDLE_NAMES} from '../component-systems/RapierPhysicsSystem'
+import { INPUT_COLLIDER_HANDLE_NAMES} from '../component-systems/RapierPhysicsSystem'
 
 const HOVER_DISTANCE = 0.05
 const PINCH_DISTANCE = 0.005
