+++ webpack.config.js
@@ -10,7 +10,7 @@ module.exports = {
   },
 
   optimization: {
-    minimize: false,
+    minimize: true,
   },
 
   experiments: {
